cv-generator-gcp/docx_to_pdf_function/requirements.txt
functions-framework==3.*
docx2pdf==0.1.8
google-cloud-storage==2.* 